Only support the new UI for all Product pages
