+++ lib/nerves_hub_web.ex
@@ -74,7 +74,7 @@ defmodule NervesHubWeb do
 
   def updated_live_view() do
     quote do
-      use NervesHubWeb.LiveView,
+      use Phoenix.LiveView,
         layout: {NervesHubWeb.LayoutView, :live},
         container: {:div, class: "h-screen"}
 
@@ -128,24 +128,16 @@ defmodule NervesHubWeb do
       end
 
       defp setup_tab_components(socket, tabs \\ []) do
-        if socket.assigns[:new_ui] do
-          Enum.reduce(tabs, socket, fn component, socket ->
-            component.connect(socket)
-          end)
-          |> put_private(:tabs, tabs)
-        else
-          socket
-        end
+        Enum.reduce(tabs, socket, fn component, socket ->
+          component.connect(socket)
+        end)
+        |> put_private(:tabs, tabs)
       end
 
       defp update_tab_component_hooks(socket) do
-        if socket.assigns[:new_ui] do
-          socket
-          |> detach_hooks()
-          |> attach_hooks()
-        else
-          socket
-        end
+        socket
+        |> detach_hooks()
+        |> attach_hooks()
       end
 
       defp detach_hooks(socket) do
