+++ lib/nerves_hub_web/components/device_page/health_tab.ex
@@ -256,7 +256,7 @@ defmodule NervesHubWeb.Components.DevicePage.HealthTab do
 
       types(charts) != types(data) ->
         push_patch(socket,
-          to: ~p"/org/#{org.name}/#{product.name}/devices/#{device.identifier}/healthz"
+          to: ~p"/org/#{org.name}/#{product.name}/devices/#{device.identifier}/health"
         )
 
       true ->
