+++ lib/nerves_hub_web/live/archive_templates/show_archive_template.html.heex
@@ -1,60 +1,118 @@
-<div class="action-row">
-  <.link patch={~p"/org/#{@org.name}/#{@product.name}/archives"} class="back-link">
-    All Archives
-  </.link>
-  <div class="btn-group" role="group" aria-label="Device Actions">
-    <.link class="btn btn-outline-light btn-action" aria-label="Download" href={~p"/org/#{@org.name}/#{@product.name}/archives/#{@archive.uuid}/download"} download>
-      <span class="button-icon download"></span>
-      <span class="action-text">Download</span>
+<div class="h-[56px] skrink-0 flex justify-between bg-base-900 border-b border-base-700 pl-6 items-center">
+  <div class="flex gap-2.5">
+    <.link navigate={~p"/org/#{@org.name}/#{@product.name}/archives"} class="back-link flex gap-2.5 items-center">
+      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20" fill="none">
+        <path d="M4.16671 10L9.16671 5M4.16671 10L9.16671 15M4.16671 10H15.8334" stroke="#A1A1AA" stroke-width="1.2" stroke-linecap="round" stroke-linejoin="round" />
+      </svg>
+      <span class="text-base-400">All Archives</span>
     </.link>
-    <.link class="btn btn-outline-light btn-action" aria-label="Delete" phx-click="delete-archive" data-confirm="Are you sure you want to delete this archive? This can not be undone.">
-      <span class="button-icon delete"></span>
-      <span class="action-text">Delete</span>
-    </.link>
+    <span class="text-base-400">/</span>
+    <span class="text-zinc-50 font-semibold">{@archive.uuid}</span>
   </div>
 </div>
 
-<h1>Archive {@archive.version}</h1>
+<div class="h-0 flex-1 overflow-y-auto flex flex-col">
+  <div class="h-[90px] shrink-0 flex justify-between p-6 border-b border-zinc-700">
+    <div class="flex gap-3 items-center">
+      <h1 class="text-xl font-semibold leading-[30px] text-zinc-50">
+        {@archive.uuid}
+      </h1>
+    </div>
 
-<div class="firmware-meta-grid">
-  <div>
-    <div class="help-text">UUID</div>
-    <p class="ff-m">{@archive.uuid}</p>
+    <div class="flex items-center gap-2">
+      <.button type="link" href={~p"/org/#{@org.name}/#{@product.name}/archives/#{@archive.uuid}/download"} download>
+        <.icon name="download" />Download
+      </.button>
+
+      <button
+        class="box-content h-5 py-1.5 px-3 flex gap-2 justify-center items-center rounded border border-red-500 bg-zinc-800 text-sm font-medium text-zinc-300 disabled:text-zinc-500"
+        aria-label="Delete archive"
+        type="button"
+        phx-click="delete-archive"
+      >
+        <svg class="size-5 stroke-red-500" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path
+            d="M12.4999 5.83337H7.49992M12.4999 5.83337H14.9999M12.4999 5.83337C12.4999 4.45266 11.3806 3.33337 9.99992 3.33337C8.61921 3.33337 7.49992 4.45266 7.49992 5.83337M7.49992 5.83337H4.99992M3.33325 5.83337H4.99992M4.99992 5.83337V15C4.99992 15.9205 5.74611 16.6667 6.66659 16.6667H13.3333C14.2537 16.6667 14.9999 15.9205 14.9999 15V5.83337M14.9999 5.83337H16.6666M8.33325 9.16671V13.3334M11.6666 13.3334V9.16671"
+            stroke-width="1.2"
+            stroke-linecap="round"
+            stroke-linejoin="round"
+          />
+        </svg>
+        <span class="text-sm font-medium text-red-500">Delete</span>
+      </button>
+    </div>
   </div>
-  <div>
-    <div class="help-text">Platform</div>
-    <p>{@archive.platform}</p>
-  </div>
-  <div>
-    <div class="help-text">Architecture</div>
-    <p>{@archive.architecture}</p>
-  </div>
-  <div class="gr-2">
-    <div class="help-text">Author</div>
-    <p>{@archive.author}</p>
-  </div>
-  <div class="gr-2">
-    <div class="help-text">Uploaded On</div>
-    <p class="date-time">
-      {@archive.inserted_at}
-    </p>
-  </div>
-  <div class="gr-2">
-    <div class="help-text">VCS ID</div>
-    <p class="ff-m">
-      <%= if is_nil(@archive.vcs_identifier) do %>
-        -
-      <% else %>
-        {@archive.vcs_identifier}
-      <% end %>
-    </p>
-  </div>
-  <div class="gr-2">
-    <div class="help-text">Signing Key</div>
-    <p class="ff-m">
-      <span class="badge">{format_signed(@archive, @org_keys)}</span>
-    </p>
+
+  <div class="h-full flex items-start justify-between gap-4 p-6">
+    <div class="w-1/2 flex flex-col gap-4">
+      <div class="flex flex-col pb-4 rounded border border-zinc-700 bg-zinc-900 shadow-device-details-content">
+        <div class="h-14 pl-4 pr-3 flex items-center text-neutral-50 font-medium leading-6">
+          General info
+        </div>
+        <div class="flex flex-col gap-3">
+          <div :if={@archive.author} class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Author:</span>
+            <span class="text-sm text-zinc-300">{@archive.author}</span>
+          </div>
+
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Description:</span>
+            <span :if={@archive.description} class="text-sm text-zinc-300">{@archive.description}</span>
+            <span :if={!@archive.description} class="text-sm text-zinc-400">No description</span>
+          </div>
+
+          <div :if={@archive.misc} class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Misc:</span>
+            <span class="text-sm text-zinc-300">{@archive.misc}</span>
+          </div>
+
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Signed with:</span>
+            <span class="flex items-center gap-1 pl-2.5 pr-3 py-0.5 border border-zinc-700 rounded-full bg-zinc-800">
+              <code class="text-xs text-zinc-300">{format_signed(@archive, @org_keys)}</code>
+            </span>
+          </div>
+
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Uploaded on:</span>
+            <span class="text-sm text-zinc-300">{Calendar.strftime(@archive.inserted_at, "%Y-%m-%d at %I:%M %p UTC")}</span>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    <div class="w-1/2 flex flex-col gap-4">
+      <div class="flex flex-col pb-4 rounded border border-zinc-700 bg-zinc-900 shadow-device-details-content">
+        <div class="h-14 pl-4 pr-3 flex items-center text-neutral-50 font-medium leading-6">
+          Metadata
+        </div>
+        <div class="flex flex-col gap-3">
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Version:</span>
+            <span class="text-sm text-zinc-300">{@archive.version}</span>
+          </div>
+
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Architecture:</span>
+            <span class="text-sm text-zinc-300">{@archive.architecture}</span>
+          </div>
+
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Platform:</span>
+            <span class="text-sm text-zinc-300">{@archive.platform}</span>
+          </div>
+
+          <div :if={@archive.vcs_identifier} class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">VCS ID:</span>
+            <span class="text-sm text-zinc-300">{@archive.vcs_identifier}</span>
+          </div>
+
+          <div class="min-h-7 px-4 flex gap-4 items-center">
+            <span class="text-sm text-nerves-gray-500">Size:</span>
+            <span class="text-sm text-zinc-300">{format_file_size(@archive.size)}</span>
+          </div>
+        </div>
+      </div>
+    </div>
   </div>
 </div>
-
-<div class="divider"></div>
