--- lib/nerves_hub_web/live/deployment_groups/newz.html.heex
@@ -1,119 +1,0 @@
-<div class="h-[56px] skrink-0 flex justify-between bg-base-900 border-b border-base-700 pl-6 items-center">
-  <div class="flex gap-2.5">
-    <.link navigate={~p"/org/#{@org.name}/#{@product.name}/deployment_groups"} class="back-link flex gap-2.5 items-center">
-      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20" fill="none">
-        <path d="M4.16671 10L9.16671 5M4.16671 10L9.16671 15M4.16671 10H15.8334" stroke="#A1A1AA" stroke-width="1.2" stroke-linecap="round" stroke-linejoin="round" />
-      </svg>
-      <span class="text-base-400">All Deployment Groups</span>
-    </.link>
-    <span class="text-base-400">/</span>
-    <span class="text-zinc-50 font-semibold">Add Deployment Group</span>
-  </div>
-
-  <div class="h-full border-l flex items-center justify-center border-base-700 bg-base-900">
-    <a :if={Application.get_env(:nerves_hub, :new_ui)} href={"/ui/switch?return_to=#{@current_path}"} class="">
-      <svg class="box-content px-5 h-5 w-5 stroke-zinc-500 hover:stroke-indigo-500" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path
-          d="M21 14V19C21 20.1046 20.1046 21 19 21H16M3 16V19C3 20.1046 3.89543 21 5 21H16M3 16V5C3 3.89543 3.89543 3 5 3H11M3 16C4.40293 15.7662 6.63687 15.7073 8.94504 16.2427M16 21C14.2965 18.2317 11.5726 16.8522 8.94504 16.2427M8.94504 16.2427C9.87157 15.1698 11.1851 14.1585 13 13.3925M8.5 7C8 7 7 7.3 7 8.5C7 9.7 8 10 8.5 10C9 10 10 9.7 10 8.5C10 7.3 9 7 8.5 7ZM17.5 9.46262L14.7188 11L15.25 7.74377L13 5.43769L16.1094 4.96262L17.5 2L18.8906 4.96262L22 5.43769L19.75 7.74377L20.2812 11L17.5 9.46262Z"
-          stroke-width="1.2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        />
-      </svg>
-    </a>
-  </div>
-</div>
-
-<div class="h-0 flex-1 overflow-y-auto">
-  <div class="flex items-center h-[90px] gap-4 px-6 py-7 border-b border-[#3F3F46] text-sm font-medium">
-    <h1 class="text-base font-semibold text-neutral-50">Add Deployment Group</h1>
-  </div>
-
-  <div :if={@firmware_required} class="h-full pb-16 flex flex-col items-center justify-center gap-4 p-6">
-    <span class="text-xl font-medium text-neutral-50">Please upload your first firmware before creating a deployment group.</span>
-    <.link class="underline underline-offset-4 decoration-dotted hover:text-neutral-50" navigate={~p"/org/#{@org.name}/#{@product.name}/firmware"} aria-label="Add a new deployment group">
-      Go to the firmware page
-    </.link>
-  </div>
-
-  <div :if={!@firmware_required} class="flex flex-col items-start justify-between gap-4 p-6">
-    <.form for={@form} class="w-full" phx-submit="create-deployment-group">
-      <div class="flex flex-col w-2/3 bg-zinc-900 border border-zinc-700 rounded">
-        <div class="flex justify-between items-center h-14 px-4 border-b border-zinc-700">
-          <div class="text-base text-neutral-50 font-medium">General settings</div>
-        </div>
-
-        <div class="flex p-6 gap-6 border-b border-zinc-700">
-          <div class="w-2/3 flex flex-col gap-6">
-            <.input field={@form[:name]} label="Name" placeholder="eg. Production" />
-          </div>
-        </div>
-
-        <div class="flex p-6 gap-6 border-b border-zinc-700">
-          <div class="w-2/3 flex flex-col gap-6">
-            <div class="flex gap-6">
-              <div class="w-1/2">
-                <.input field={@form[:platform]} type="select" options={@platforms} phx-change="platform-selected" label="Platform" prompt="Choose a platform" />
-              </div>
-
-              <div class="w-1/2">
-                <.input
-                  field={@form[:architecture]}
-                  type="select"
-                  options={@architectures}
-                  disabled={@architectures == []}
-                  phx-change="architecture-selected"
-                  label="Architecture"
-                  prompt={(@architectures != [] && "Choose an architecture") || ""}
-                />
-              </div>
-            </div>
-
-            <.input
-              field={@form[:firmware_id]}
-              type="select"
-              options={firmware_dropdown_options(@firmwares)}
-              disabled={@firmwares == []}
-              label="Firmware"
-              prompt={(@firmwares != [] && "Choose a firmware") || "Please select a platform and architecture first"}
-              hint="This firmware will be distributed to devices assigned to this deployment group."
-            />
-
-            <%!-- <.input field={@form[:tags]} label="Description" placeholder="eg. sensor hub at customer X" />
-
-            <.input field={@form[:version]} value={tags_to_string(@form[:tags])} label="Tags" placeholder="eg. batch-123" /> --%>
-          </div>
-        </div>
-
-        <div class="flex p-6 gap-6">
-          <div class="w-2/3 flex flex-col gap-6">
-            <div class="text-base text-neutral-50 font-medium">Device matching conditions</div>
-            <div class="flex flex-col gap-3">
-              <p class="text-sm text-zinc-400">
-                These conditions are used for matching devices which don't have a configured deployment group.
-              </p>
-              <p class="text-sm text-zinc-400">
-                The matching is undertaken when a device connects to the platform.
-              </p>
-            </div>
-            <div class="flex gap-6">
-              <div class="w-1/2">
-                <.input field={@form[:tags]} value={tags_to_string(@form[:conditions])} label="Tag(s) distributed to" placeholder="eg. batch-123" />
-              </div>
-
-              <div class="w-1/2">
-                <.input field={@form[:version]} value={@form[:conditions].value["version"]} label="Version requirement" placeholder="eg. 1.2.3" />
-              </div>
-            </div>
-          </div>
-        </div>
-
-        <div class="flex justify-between items-center h-14 px-4 border-t border-zinc-700">
-          <.button style="secondary" type="submit">
-            <.icon name="save" /> Save changes
-          </.button>
-        </div>
-      </div>
-    </.form>
-  </div>
-</div>
