--- lib/nerves_hub_web/live/deployment_groups/show-new.html.heex
@@ -1,111 +1,0 @@
-<div class="h-[56px] skrink-0 flex justify-end bg-base-900 border-b border-base-700 pl-6 items-center">
-  <div class="flex gap-2.5">
-    <.link navigate={~p"/org/#{@org.name}/#{@product.name}/deployment_groups"} class="back-link flex gap-2.5 items-center">
-      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20" fill="none">
-        <path d="M4.16671 10L9.16671 5M4.16671 10L9.16671 15M4.16671 10H15.8334" stroke="#A1A1AA" stroke-width="1.2" stroke-linecap="round" stroke-linejoin="round" />
-      </svg>
-      <span class="text-base-400">All Deployment Groups</span>
-    </.link>
-    <span class="text-base-400">/</span>
-    <span class="text-zinc-50 font-semibold">{@deployment_group.name}</span>
-  </div>
-  <div class="flex items-center gap-2 ml-auto mr-6">
-    <.button
-      :if={authorized?(:"deployment_group:toggle", @org_user)}
-      aria-label={"#{(@deployment_group.is_active && "Pause") || "Resume"} deployment"}
-      type="button"
-      phx-click="toggle"
-      data-confirm={"Are you sure you want to #{(@deployment_group.is_active && "pause") || "resume"} the deployment?"}
-    >
-      <.icon name="power" />{(@deployment_group.is_active && "Pause") || "Resume"}
-    </.button>
-  </div>
-  <div class="h-full border-l flex items-center justify-center border-base-700 bg-base-900">
-    <a :if={Application.get_env(:nerves_hub, :new_ui)} href={"/ui/switch?return_to=#{@current_path}"} class="">
-      <svg class="box-content px-5 h-5 w-5 stroke-zinc-500 hover:stroke-indigo-500" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path
-          d="M21 14V19C21 20.1046 20.1046 21 19 21H16M3 16V19C3 20.1046 3.89543 21 5 21H16M3 16V5C3 3.89543 3.89543 3 5 3H11M3 16C4.40293 15.7662 6.63687 15.7073 8.94504 16.2427M16 21C14.2965 18.2317 11.5726 16.8522 8.94504 16.2427M8.94504 16.2427C9.87157 15.1698 11.1851 14.1585 13 13.3925M8.5 7C8 7 7 7.3 7 8.5C7 9.7 8 10 8.5 10C9 10 10 9.7 10 8.5C10 7.3 9 7 8.5 7ZM17.5 9.46262L14.7188 11L15.25 7.74377L13 5.43769L16.1094 4.96262L17.5 2L18.8906 4.96262L22 5.43769L19.75 7.74377L20.2812 11L17.5 9.46262Z"
-          stroke-width="1.2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-        />
-      </svg>
-    </a>
-  </div>
-</div>
-
-<div class="h-0 flex-1 overflow-y-auto flex flex-col">
-  <div class="h-[90px] shrink-0 flex justify-between p-6">
-    <div class="flex gap-3 items-center">
-      <svg class={["size-3", (@deployment_group.is_active && "fill-emerald-500") || "fill-zinc-500"]} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 6 6">
-        <circle cx="3" cy="3" r="3" />
-      </svg>
-      <h1 class="text-xl font-semibold leading-[30px] text-zinc-50">
-        {@deployment_group.name}
-      </h1>
-    </div>
-
-    <div class="flex items-center gap-2">
-      <div class="flex h-7 py-1 px-2 items-center rounded bg-zinc-800">
-        <span class="text-sm text-zinc-400 mr-1">Devices:</span>
-
-        <div class="flex items-center">
-          <span class="text-sm text-base-300 mr-1">{@deployment_group.device_count || 0}</span>
-        </div>
-      </div>
-      <div class="flex h-7 py-1 px-2 items-center rounded bg-zinc-800">
-        <span class="text-sm text-zinc-400 mr-1">Arch:</span>
-
-        <div class="flex items-center">
-          <span class="text-sm text-base-300 mr-1">{@deployment_group.firmware.architecture}</span>
-        </div>
-      </div>
-
-      <div class="flex h-7 py-1 px-2 items-center rounded bg-zinc-800">
-        <span class="text-sm text-zinc-400 mr-1">Platform:</span>
-
-        <div class="flex items-center">
-          <span class="text-sm text-base-300 mr-1">{@deployment_group.firmware.platform}</span>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="flex w-full justify-between px-6 border-b border-zinc-700">
-    <div class="flex">
-      <.link class={tab_classes(@tab, :summary)} patch={~p"/org/#{@org.name}/#{@product.name}/deployment_groups/#{@deployment_group.name}"}>
-        Summary
-      </.link>
-      <.link class={tab_classes(@tab, :release_history)} patch={~p"/org/#{@org.name}/#{@product.name}/deployment_groups/#{@deployment_group.name}/releases"}>
-        Release History
-      </.link>
-      <.link class={tab_classes(@tab, :activity)} patch={~p"/org/#{@org.name}/#{@product.name}/deployment_groups/#{@deployment_group.name}/activity"}>
-        Activity
-      </.link>
-      <.link class={tab_classes(@tab, :settings)} patch={~p"/org/#{@org.name}/#{@product.name}/deployment_groups/#{@deployment_group.name}/settings"}>
-        Settings
-      </.link>
-    </div>
-  </div>
-
-  <.live_component
-    :if={@tab == :summary}
-    module={SummaryTab}
-    id="deployment_group_summary"
-    deployment_group={@deployment_group}
-    inflight_updates={@inflight_updates}
-    up_to_date_count={@up_to_date_count}
-    waiting_for_update_count={@waiting_for_update_count}
-    updating_count={@updating_count}
-    product={@product}
-    org={@org}
-    org_user={@org_user}
-    user={@user}
-  />
-
-  <.live_component :if={@tab == :release_history} module={ReleaseHistoryTab} id="deployment_group_release_history" />
-
-  <.live_component :if={@tab == :activity} module={ActivityTab} id="deployment_group_activity" deployment_group={@deployment_group} org={@org} product={@product} user={@user} />
-
-  <.live_component :if={@tab == :settings} module={SettingsTab} id="deployment_group_settings" deployment_group={@deployment_group} org={@org} product={@product} user={@user} org_user={@org_user} />
-</div>
