--- lib/nerves_hub_web/live/devices/device_health.html.heex
@@ -1,65 +1,0 @@
-<div class="action-row mb-1">
-  <.link navigate={~p"/org/#{@org.name}/#{@product.name}/devices/#{@device.identifier}"} class="back-link">Back to
-    device </.link>
-</div>
-<HealthHeader.render
-  org_name={@org.name}
-  product_name={@product.name}
-  device={@device}
-  latest_connection={@latest_connection}
-  health_reported_at={@latest_metrics["timestamp"]}
-  health_check_timer={@health_check_timer}
-/>
-
-<div class="divider"></div>
-
-<div class="action-row mb-2">
-  <h3 class="mb-2">Device Metrics</h3>
-
-  <div class="btn-group metrics-btn-group" role="group" aria-label="Metrics Time Interval">
-    <%= for {unit, amount} <- @time_frame_opts do %>
-      <button
-        class={"btn btn-outline-light btn-action #{if {unit, amount}==@time_frame, do: "metrics-active" }"}
-        aria-label="1 Hour"
-        type="button"
-        phx-click="set-time-frame"
-        phx-value-unit={unit}
-        phx-value-amount={amount}
-      >
-        <span class="action-text">
-          {Integer.to_string(amount) <> " " <> unit <> if amount > 1, do: "s", else: ""}
-        </span>
-      </button>
-    <% end %>
-  </div>
-</div>
-
-<div class="action">
-  <%= if @charts==[] do %>
-    <div class="col-lg-16">
-      <p>No data for selected period.</p>
-    </div>
-  <% else %>
-    <div class="col-lg-16">
-      <%= for chart <- @charts do %>
-        <div style="margin-bottom: 100px;">
-          <canvas
-            id={chart.type}
-            style="display: block; box-sizing: border-box;"
-            width="1200"
-            height="400"
-            phx-hook="Chart"
-            phx-update="ignore"
-            data-type={Jason.encode!(chart.type)}
-            data-unit={Jason.encode!(chart.unit)}
-            data-max={Jason.encode!(chart.max)}
-            data-min={Jason.encode!(chart.min)}
-            data-metrics={Jason.encode!(chart.data)}
-            data-title={Jason.encode!(chart.title)}
-          >
-          </canvas>
-        </div>
-      <% end %>
-    </div>
-  <% end %>
-</div>
