+++ lib/nerves_hub_web/live/devices/show.ex
@@ -3,39 +3,25 @@ defmodule NervesHubWeb.Live.Devices.Show do
 
   require Logger
 
-  alias NervesHub.AuditLogs
   alias NervesHub.AuditLogs.DeviceTemplates
   alias NervesHub.Devices
-  alias NervesHub.Devices.Alarms
   alias NervesHub.Devices.Connections
-  alias NervesHub.Devices.Metrics
-  alias NervesHub.Devices.UpdatePayload
   alias NervesHub.Extensions.Health
-  alias NervesHub.Firmwares
-  alias NervesHub.ManagedDeployments
   alias NervesHub.Repo
-  alias NervesHub.Scripts
-  alias NervesHub.Tracker
-
-  alias NervesHubWeb.Components.AuditLogFeed
-  alias NervesHubWeb.Components.DeviceHeader
-  alias NervesHubWeb.Components.DeviceLocation
-  alias NervesHubWeb.Components.DeviceUpdateStatus
-  alias NervesHubWeb.Components.FwupProgress
-  alias NervesHubWeb.Components.Utils
 
   alias NervesHubWeb.Components.DevicePage.ActivityTab
   alias NervesHubWeb.Components.DevicePage.ConsoleTab
   alias NervesHubWeb.Components.DevicePage.DetailsTab
   alias NervesHubWeb.Components.DevicePage.HealthTab
   alias NervesHubWeb.Components.DevicePage.SettingsTab
-
-  @tab_components [ActivityTab, ConsoleTab, DetailsTab, HealthTab, SettingsTab]
+  alias NervesHubWeb.Components.DeviceUpdateStatus
+  alias NervesHubWeb.Components.FwupProgress
 
   alias NervesHubWeb.Presence
+
   alias Phoenix.Socket.Broadcast
 
-  @running_script_placeholder "Running Script.."
+  @tab_components [ActivityTab, ConsoleTab, DetailsTab, HealthTab, SettingsTab]
 
   def mount(%{"device_identifier" => device_identifier}, _session, socket) do
     %{org: org, product: product, user: user} = socket.assigns
@@ -55,16 +41,10 @@ defmodule NervesHubWeb.Live.Devices.Show do
     |> sidebar_tab(:devices)
     |> selected_tab()
     |> assign(:extension_overrides, extension_overrides(device, product))
-    |> assign(:scripts, scripts_with_output(product))
     |> general_assigns(device)
-    |> assign_metadata()
     |> schedule_health_check_timer()
     |> assign(:fwup_progress, nil)
-    |> assign(:page_number, 1)
-    |> assign(:page_size, 5)
     |> assign(:pinned?, Devices.device_pinned?(user.id, device.id))
-    |> audit_log_assigns()
-    |> assign_deployment_groups()
     |> setup_presence_tracking()
     |> setup_tab_components(@tab_components)
     |> ok()
@@ -96,13 +76,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     else
       {:noreply, stream_insert(socket, :presences, presence)}
     end
-  end
-
-  # can be removed when the old UI is removed
-  def handle_info(%Broadcast{topic: "firmware", event: "created"}, socket) do
-    firmware = Firmwares.get_firmware_for_device(socket.assigns.device)
-
-    {:noreply, assign(socket, :firmwares, firmware)}
   end
 
   def handle_info(%Broadcast{event: "connection:heartbeat"}, socket) do
@@ -144,7 +117,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     socket
     |> assign(:device, device)
     |> assign(:device_connection, device.latest_connection)
-    |> assign(:console_active?, Tracker.console_active?(device))
     |> assign(:fwup_progress, nil)
     |> assign(:update_information, Devices.resolve_update(device))
     |> then(fn socket ->
@@ -153,34 +125,15 @@ defmodule NervesHubWeb.Live.Devices.Show do
     |> noreply()
   end
 
-  def handle_info(%Broadcast{event: "console_joined"}, socket) do
-    socket
-    |> assign(:console_active?, true)
-    |> noreply()
-  end
-
   def handle_info(%Broadcast{event: "fwup_progress", payload: payload}, socket) do
     if payload.percent == 100 do
       socket
-      |> put_flash(:info, "Update complete: The device will reboot shortly.")
       |> send_toast(:info, "Update complete: The device will reboot shortly.")
       |> assign(:fwup_progress, nil)
       |> noreply()
     else
       {:noreply, assign(socket, :fwup_progress, payload.percent)}
     end
-  end
-
-  def handle_info(
-        %Broadcast{event: "health_check_report"},
-        %{assigns: %{device: device}} = socket
-      ) do
-    latest_metrics = Metrics.get_latest_metric_set(device.id)
-
-    socket
-    |> assign(:latest_metrics, latest_metrics)
-    |> assign_metadata()
-    |> noreply
   end
 
   def handle_info(:check_health_interval, socket) do
@@ -191,14 +144,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     socket
     |> assign(:health_check_timer, timer_ref)
     |> noreply()
-  end
-
-  def handle_info(%Broadcast{event: "location:updated"}, socket) do
-    %{device: device, org: org} = socket.assigns
-
-    {:ok, device} = Devices.get_device_by_identifier(org, device.identifier, :latest_connection)
-
-    {:noreply, assign(socket, :device, device)}
   end
 
   # Ignore unknown messages
@@ -278,27 +223,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     {:noreply, put_flash(socket, :info, "Device identification requested")}
   end
 
-  # TODO: [OLD UI] Can we removed when we remove the old UI
-  def handle_event("toggle-health-check-auto-refresh", _value, socket) do
-    if timer_ref = socket.assigns.health_check_timer do
-      _ = Process.cancel_timer(timer_ref)
-      {:noreply, assign(socket, :health_check_timer, nil)}
-    else
-      {:noreply, schedule_health_check_timer(socket)}
-    end
-  end
-
-  def handle_event("paginate", %{"page" => page_num}, socket) do
-    params = %{"page_size" => socket.assigns.page_size, "page_number" => page_num}
-
-    url =
-      ~p"/org/#{socket.assigns.org.name}/#{socket.assigns.product.name}/devices/#{socket.assigns.device.identifier}/activity?#{params}"
-
-    socket
-    |> push_patch(to: url)
-    |> noreply()
-  end
-
   def handle_event("clear-penalty-box", _params, socket) do
     %{org_user: org_user, user: user, device: device} = socket.assigns
 
@@ -309,210 +233,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     socket
     |> assign(:device, updated_device)
     |> send_toast(:info, "Device removed from the penalty box, and firmware updates enabled.")
-    |> noreply()
-  end
-
-  def handle_event("toggle-deployment-firmware-updates", _params, socket) do
-    %{org_user: org_user, user: user, device: device} = socket.assigns
-
-    authorized!(:"device:toggle-updates", org_user)
-
-    {:ok, updated_device} = Devices.toggle_automatic_updates(device, user)
-
-    message = [
-      "Firmware updates ",
-      (updated_device.updates_enabled && "enabled") || "disabled",
-      "."
-    ]
-
-    socket
-    |> assign(:device, updated_device)
-    |> send_toast(:info, Enum.join(message))
-    |> noreply()
-  end
-
-  def handle_event("restore", _, socket) do
-    authorized!(:"device:restore", socket.assigns.org_user)
-
-    {:ok, device} = Devices.restore_device(socket.assigns.device)
-
-    {:noreply, assign(socket, :device, device)}
-  end
-
-  def handle_event("destroy", _, socket) do
-    %{org: org, org_user: org_user, product: product, device: device} = socket.assigns
-
-    authorized!(:"device:destroy", org_user)
-
-    {:ok, _device} = Devices.destroy_device(device)
-
-    socket
-    |> put_flash(:info, "Device destroyed successfully.")
-    |> push_navigate(to: ~p"/org/#{org.name}/#{product.name}/devices")
-    |> noreply()
-  end
-
-  def handle_event("delete", _, socket) do
-    authorized!(:"device:delete", socket.assigns.org_user)
-
-    {:ok, device} = Devices.delete_device(socket.assigns.device)
-
-    {:noreply, assign(socket, :device, device)}
-  end
-
-  def handle_event("set-deployment-group", %{"deployment_id" => ""}, socket) do
-    socket
-    |> put_flash(:error, "Please select a deployment group.")
-    |> noreply()
-  end
-
-  def handle_event(
-        "set-deployment-group",
-        %{"deployment_id" => deployment_id},
-        %{assigns: %{user: user, device: device, deployment_groups: deployment_groups}} =
-          socket
-      ) do
-    deployment_group = Enum.find(deployment_groups, &(&1.id == String.to_integer(deployment_id)))
-    device = Devices.update_deployment_group(device, deployment_group)
-    _ = DeviceTemplates.audit_device_deployment_group_update(user, device, deployment_group)
-
-    socket
-    |> assign(:device, device)
-    |> assign(:deployment_group, deployment_group)
-    |> put_flash(:info, "Deployment Group successfully updated")
-    |> noreply()
-  end
-
-  def handle_event("push-update", %{"uuid" => ""}, socket) do
-    socket
-    |> put_flash(:error, "Please select a firmware you would like to send to the device.")
-    |> noreply()
-  end
-
-  def handle_event("push-update", %{"uuid" => uuid}, socket) do
-    authorized!(:"device:push-update", socket.assigns.org_user)
-
-    %{product: product, device: device, user: user} = socket.assigns
-
-    {:ok, firmware} = Firmwares.get_firmware_by_product_and_uuid(product, uuid)
-    {:ok, url} = Firmwares.get_firmware_url(firmware)
-    {:ok, meta} = Firmwares.metadata_from_firmware(firmware)
-    {:ok, device} = Devices.disable_updates(device, user)
-
-    DeviceTemplates.audit_firmware_pushed(user, device, firmware)
-
-    payload = %UpdatePayload{
-      update_available: true,
-      firmware_url: url,
-      firmware_meta: meta
-    }
-
-    _ = NervesHubWeb.Endpoint.broadcast("device:#{device.id}", "devices/update-manual", payload)
-
-    socket
-    |> assign(:device, device)
-    |> put_flash(:info, "Pushing firmware update")
-    |> noreply()
-  end
-
-  def handle_event("push-available-update", _, socket) do
-    authorized!(:"device:push-update", socket.assigns.org_user)
-
-    %{device: device, deployment_group: deployment_group, user: user} = socket.assigns
-
-    deployment_group = NervesHub.Repo.preload(deployment_group, :firmware)
-
-    case Devices.told_to_update(device, deployment_group) do
-      {:ok, _inflight_update} ->
-        DeviceTemplates.audit_pushed_available_update(user, device, deployment_group)
-
-        socket
-        |> put_flash(:info, "Pushing available firmware update")
-        |> noreply()
-
-      :error ->
-        Logger.error(
-          "An inflight update could not be created or found for the device #{device.identifier} (#{device.id})"
-        )
-
-        socket
-        |> put_flash(
-          :info,
-          "There was an error sending the update to the device. Please contact support."
-        )
-        |> noreply()
-    end
-  end
-
-  def handle_event(
-        "run-script",
-        %{"idx" => index},
-        %{assigns: %{device: device, scripts: scripts, org_user: org_user}} = socket
-      ) do
-    authorized!(:"support_script:run", org_user)
-
-    {script, idx} = Enum.at(scripts, String.to_integer(index))
-
-    socket
-    |> assign(:scripts, update_script_output(scripts, idx, @running_script_placeholder))
-    |> start_async({:run_script, idx}, fn -> Scripts.Runner.send(device, script) end)
-    |> noreply()
-  end
-
-  def handle_event(
-        "clear-script-output",
-        %{"idx" => index},
-        %{assigns: %{scripts: scripts}} = socket
-      ) do
-    socket
-    |> assign(:scripts, update_script_output(scripts, String.to_integer(index), nil))
-    |> noreply()
-  end
-
-  def handle_event(
-        "remove-from-deployment-group",
-        _,
-        %{assigns: %{device: device}} = socket
-      ) do
-    device =
-      device
-      |> Devices.clear_deployment_group()
-      |> Repo.preload(:deployment_group)
-
-    socket
-    |> assign(:device, device)
-    |> assign(:deployment_group, nil)
-    |> assign_deployment_groups()
-    |> put_flash(:info, "Device successfully removed from the deployment")
-    |> noreply()
-  end
-
-  def handle_event("set-paginate-opts", %{"page-size" => page_size}, socket) do
-    params = %{"page_size" => page_size, "page_number" => 1}
-
-    url =
-      ~p"/org/#{socket.assigns.org.name}/#{socket.assigns.product.name}/devices/#{socket.assigns.device.identifier}/activity?#{params}"
-
-    socket
-    |> push_patch(to: url)
-    |> noreply()
-  end
-
-  def handle_async({:run_script, index}, result, %{assigns: %{scripts: scripts}} = socket) do
-    output =
-      case result do
-        {:ok, {:ok, output}} ->
-          output
-
-        {:ok, {:error, reason}} ->
-          "Error: #{reason}"
-
-        e ->
-          inspect(e)
-      end
-
-    socket
-    |> assign(:scripts, update_script_output(scripts, index, output))
     |> noreply()
   end
 
@@ -530,44 +250,13 @@ defmodule NervesHubWeb.Live.Devices.Show do
     socket = stream(socket, :presences, [])
 
     if connected?(socket) do
-      Presence.track_user(topic, user.id, %{name: user.name})
-      Presence.subscribe(topic)
+      {:ok, _} = Presence.track_user(topic, user.id, %{name: user.name})
+      :ok = Presence.subscribe(topic)
       stream(socket, :presences, Presence.list_online_users(topic))
     else
       socket
     end
   end
-
-  defp scripts_with_output(product) do
-    product
-    |> Scripts.all_by_product()
-    |> Enum.map(&Map.put(&1, :output, nil))
-    |> Enum.with_index()
-  end
-
-  defp update_script_output(scripts, index, output) do
-    List.update_at(scripts, index, fn {script, idx} ->
-      {%{script | output: output}, idx}
-    end)
-  end
-
-  defp assign_metadata(%{assigns: %{device: device}} = socket) do
-    health = Devices.get_latest_health(device.id)
-
-    metadata =
-      if health, do: health.data["metadata"] || %{}, else: %{}
-
-    socket
-    |> assign(:metadata, Map.drop(metadata, standard_keys(device)))
-  end
-
-  defp standard_keys(%{firmware_metadata: nil}), do: []
-
-  defp standard_keys(%{firmware_metadata: firmware_metadata}),
-    do:
-      firmware_metadata
-      |> Map.keys()
-      |> Enum.map(&to_string/1)
 
   defp schedule_health_check_timer(socket) do
     %{device: device, product: product} = socket.assigns
@@ -584,49 +273,6 @@ defmodule NervesHubWeb.Live.Devices.Show do
     product.extensions.health and device.extensions.health
   end
 
-  defp audit_log_assigns(
-         %{assigns: %{device: device, page_number: page_number, page_size: page_size}} = socket
-       ) do
-    {logs, audit_pager} =
-      AuditLogs.logs_for_feed(device, %{page: page_number, page_size: page_size})
-
-    audit_pager = Map.from_struct(audit_pager)
-
-    socket
-    |> assign(:audit_logs, logs)
-    |> assign(:audit_pager, audit_pager)
-  end
-
-  defp assign_deployment_groups(%{assigns: %{device: %{status: :provisioned} = device}} = socket),
-    do: assign(socket, deployment_groups: ManagedDeployments.eligible_deployment_groups(device))
-
-  defp assign_deployment_groups(%{assigns: %{product: product}} = socket),
-    do:
-      assign(socket,
-        deployment_groups: ManagedDeployments.get_deployment_groups_by_product(product)
-      )
-
-  defp connecting_code(device) do
-    if device.deployment_group && device.deployment_group.connecting_code do
-      """
-      #{device.deployment_group.connecting_code}
-      #{device.connecting_code}
-      """
-    else
-      device.connecting_code
-    end
-  end
-
-  defp has_description?(description) do
-    is_binary(description) and byte_size(description) > 0
-  end
-
-  defp format_key(key) do
-    key
-    |> String.replace("_", " ")
-    |> String.capitalize()
-  end
-
   defp extension_overrides(device, product) do
     device.extensions
     |> Map.from_struct()
@@ -636,31 +282,16 @@ defmodule NervesHubWeb.Live.Devices.Show do
     |> Enum.map(&elem(&1, 0))
   end
 
-  defp running_script_placeholder(), do: @running_script_placeholder
-
-  defp script_button_text(output) when output == @running_script_placeholder or is_nil(output),
-    do: "Run"
-
-  defp script_button_text(_), do: "Close"
-
   defp disconnected?(connection) do
     is_nil(connection) || connection.status != :connected
   end
 
   defp general_assigns(socket, device) do
     socket
-    |> assign(:console_active?, Tracker.console_active?(device))
-    |> assign(:update_information, Devices.resolve_update(device))
-    |> assign(:firmwares, Firmwares.get_firmware_for_device(device))
-    |> assign(:alarms, Alarms.get_current_alarms_for_device(device))
-    |> assign(:latest_metrics, Metrics.get_latest_metric_set(device.id))
     |> assign(:deployment_group, device.deployment_group)
     |> assign(:device_connection, device.latest_connection)
     |> assign(:device, device)
   end
-
-  defp fetch_location(nil), do: %{}
-  defp fetch_location(connection), do: connection.metadata["location"]
 
   def show_menu(id, js \\ %JS{}) do
     JS.show(js, transition: "fade-in", to: "##{id}")
