--- lib/nerves_hub_web/live/firmware_templates/upload_firmware_template.html.heex
@@ -1,26 +1,0 @@
-<h1>Add Firmware</h1>
-
-<label for={@uploads.firmware.ref}>Upload a firmware file (.fw)</label>
-
-<form phx-change="firmware-selected">
-  <div class="form-group custom-upload-group" phx-drop-target={@uploads.firmware.ref}>
-    <label for={@uploads.firmware.ref} class="custom-upload-label not-selected">Click to upload file, or drag your file here.</label>
-
-    <.live_file_input upload={@uploads.firmware} required />
-
-    <div class="has-error">
-      <span :if={@error_message} class="help-block">
-        <%= case @error_code do %>
-          <% :duplicate_firmware -> %>
-            {@error_message} <a href={"/org/#{@org.name}/#{@product.name}/firmware/#{@uuid}"} target="_blank">View Firmware</a>
-          <% _ -> %>
-            {@error_message}
-        <% end %>
-      </span>
-    </div>
-
-    <div :for={entry <- @uploads.firmware.entries} class="mt-1">
-      Uploading: {entry.progress}%
-    </div>
-  </div>
-</form>
