+++ lib/nerves_hub_web/router.ex
@@ -15,20 +15,6 @@ defmodule NervesHubWeb.Router do
     plug(NervesHubWeb.Plugs.SetLocale)
   end
 
-  pipeline :dynamic_layout do
-    plug(:put_dynamic_root_layout)
-  end
-
-  def put_dynamic_root_layout(conn, _) do
-    session = Plug.Conn.get_session(conn)
-
-    if Application.get_env(:nerves_hub, :new_ui) && session["new_ui"] do
-      put_root_layout(conn, html: {NervesHubWeb.Layouts, :root})
-    else
-      put_root_layout(conn, html: {NervesHubWeb.LayoutView, :root})
-    end
-  end
-
   pipeline :logged_in do
     plug(NervesHubWeb.Plugs.FetchUser)
     plug(NervesHubWeb.Plugs.EnsureLoggedIn)
@@ -206,7 +192,6 @@ defmodule NervesHubWeb.Router do
     get("/devices/export", ProductController, :devices_export)
 
     scope "/devices/:identifier" do
-      get("/console", DeviceController, :console)
       get("/certificate/:serial/download", DeviceController, :download_certificate)
       get("/audit_logs/download", DeviceController, :export_audit_logs)
     end
@@ -222,16 +207,16 @@ defmodule NervesHubWeb.Router do
   end
 
   scope "/", NervesHubWeb do
-    pipe_through([:browser, :live_logged_in, :dynamic_layout])
+    pipe_through([:browser, :live_logged_in])
 
     get("/ui/switch", UiSwitcherController, :index)
 
     live_session :account,
+      # layout: {NervesHubWeb.Layouts, :no_sidebar},
       on_mount: [
         NervesHubWeb.Mounts.AccountAuth,
         NervesHubWeb.Mounts.EnrichSentryContext,
         NervesHubWeb.Mounts.CurrentPath,
-        {NervesHubWeb.Mounts.LayoutSelector, :no_sidebar}
       ] do
       live("/account", Live.Account, :edit)
       live("/account/delete", Live.Account, :delete)
@@ -243,13 +228,13 @@ defmodule NervesHubWeb.Router do
     end
 
     live_session :org,
+      # layout: {NervesHubWeb.Layouts, :no_sidebar},
       on_mount: [
         NervesHubWeb.Mounts.AccountAuth,
         NervesHubWeb.Mounts.EnrichSentryContext,
         NervesHubWeb.Mounts.CurrentPath,
         NervesHubWeb.Mounts.FetchOrg,
         NervesHubWeb.Mounts.FetchOrgUser,
-        {NervesHubWeb.Mounts.LayoutSelector, :no_sidebar}
       ] do
       live("/org/:org_name", Live.Org.Products, :index)
       live("/org/:org_name/new", Live.Org.Products, :new)
@@ -271,14 +256,16 @@ defmodule NervesHubWeb.Router do
     end
 
     live_session :product,
+      # The new UI is the default and only layout for the product pages.
+      root_layout: {NervesHubWeb.Layouts, :root},
+      layout: {NervesHubWeb.Layouts, :sidebar},
       on_mount: [
         NervesHubWeb.Mounts.AccountAuth,
         NervesHubWeb.Mounts.EnrichSentryContext,
         NervesHubWeb.Mounts.CurrentPath,
         NervesHubWeb.Mounts.FetchOrg,
         NervesHubWeb.Mounts.FetchOrgUser,
-        NervesHubWeb.Mounts.FetchProduct,
-        {NervesHubWeb.Mounts.LayoutSelector, :sidebar}
+        NervesHubWeb.Mounts.FetchProduct
       ] do
       live("/org/:org_name/:product_name/dashboard", Live.Dashboard.Index)
 
@@ -287,7 +274,7 @@ defmodule NervesHubWeb.Router do
       live("/org/:org_name/:product_name/devices/:device_identifier", Live.Devices.Show, :details)
 
       live(
-        "/org/:org_name/:product_name/devices/:device_identifier/healthz",
+        "/org/:org_name/:product_name/devices/:device_identifier/health",
         Live.Devices.Show,
         :health
       )
@@ -299,38 +286,25 @@ defmodule NervesHubWeb.Router do
       )
 
       live(
-        "/org/:org_name/:product_name/devices/:device_identifier/conzole",
+        "/org/:org_name/:product_name/devices/:device_identifier/console",
         Live.Devices.Show,
         :console
       )
 
       live(
-        "/org/:org_name/:product_name/devices/:device_identifier/settingz",
+        "/org/:org_name/:product_name/devices/:device_identifier/settings",
         Live.Devices.Show,
         :settings
       )
 
-      live(
-        "/org/:org_name/:product_name/devices/:device_identifier/health",
-        Live.Devices.DeviceHealth
-      )
-
-      live(
-        "/org/:org_name/:product_name/devices/:device_identifier/settings",
-        Live.Devices.Settings
-      )
-
       live("/org/:org_name/:product_name/firmware", Live.Firmware, :index)
-      live("/org/:org_name/:product_name/firmware/upload", Live.Firmware, :upload)
       live("/org/:org_name/:product_name/firmware/:firmware_uuid", Live.Firmware, :show)
 
       live("/org/:org_name/:product_name/archives", Live.Archives, :index)
-      live("/org/:org_name/:product_name/archives/upload", Live.Archives, :upload)
       live("/org/:org_name/:product_name/archives/:archive_uuid", Live.Archives, :show)
 
       live("/org/:org_name/:product_name/deployment_groups", Live.DeploymentGroups.Index)
       live("/org/:org_name/:product_name/deployments/new", Live.DeploymentGroups.New)
-      live("/org/:org_name/:product_name/deployments/newz", Live.DeploymentGroups.Newz)
 
       live(
         "/org/:org_name/:product_name/deployment_groups/:name",
@@ -354,11 +328,6 @@ defmodule NervesHubWeb.Router do
         "/org/:org_name/:product_name/deployment_groups/:name/settings",
         Live.DeploymentGroups.Show,
         :settings
-      )
-
-      live(
-        "/org/:org_name/:product_name/deployment_groups/:name/edit",
-        Live.DeploymentGroups.Edit
       )
 
       live("/org/:org_name/:product_name/scripts", Live.SupportScripts.Index)
