+++ test/nerves_hub_web/live/devices/new_ui/show_test.exs
@@ -109,11 +109,10 @@ defmodule NervesHubWeb.Live.Devices.NewUI.ShowTest do
 
     defp user_initials(user) do
       String.split(user.name)
-      |> Enum.map(fn w ->
+      |> Enum.map_join(fn w ->
         String.at(w, 0)
         |> String.upcase()
       end)
-      |> Enum.join("")
     end
   end
